nl.koenvh.screendeck.sdPlugin/code/requirements.txt
streamdeck-sdk>=1.2.0
opencv-python
streamlink
ffpyplayer
pillow